Safari-609.1.11.1/LayoutTests/fast/history/page-cache-clearing-expected.txt

Tests that clearing the page cache doesn't leave pages in the cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS window.internals.backForwardCacheSize() > 0 is true
window.internals.clearBackForwardCache()
PASS window.internals.backForwardCacheSize() is 0
PASS successfullyParsed is true

TEST COMPLETE
